+++ jdbc/src/main/java/nextstep/jdbc/core/JdbcTemplate.java
@@ -1,5 +1,6 @@
 package nextstep.jdbc.core;
 
+import nextstep.jdbc.exception.DataAccessException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
