+++ jdbc/src/main/java/nextstep/jdbc/exception/DataAccessException.java
@@ -1,4 +1,4 @@
-package nextstep.jdbc.core;
+package nextstep.jdbc.exception;
 
 public class DataAccessException extends RuntimeException {
 }
